Add documentation + MSRV + reorganize types
