+++ Cargo.toml
@@ -2,6 +2,7 @@
 name = "tiny-orm"
 version = "0.2.0"
 edition = "2021"
+rust-version = "1.74.1"
 authors = ["Matt Delacour - mdelacour.com"]
 description = "A tiny ORM for CRUD operations built on top of SQLx QueryBuilder."
 license = "MIT"
